+++ src/cmd/system/init_cli/main.go
@@ -1,0 +1,39 @@
+package main
+
+import (
+	"fmt"
+	"sm-box/internal/system/init_cli"
+	"sm-box/pkg/core/components/tracer"
+	"sm-box/pkg/core/env"
+	env_mode "sm-box/pkg/core/env/mode"
+)
+
+func init() {
+	env.Vars.SystemName = "init-script"
+	env.Version = "24.0.18"
+
+	if env.Mode == env_mode.Dev {
+		if err := tracer.Init(); err != nil {
+			panic(fmt.Sprintf("An error occurred during initialization of the function/method call trace logging component: '%s'. ", err))
+		}
+	}
+}
+
+func main() {
+	var (
+		cli init_cli.CLI
+		err error
+	)
+
+	if cli, err = init_cli.New(); err != nil {
+		panic(fmt.Sprintf("An error occurred during the creation of the '%s' instance: '%s'. ",
+			env.Vars.SystemName,
+			err))
+	}
+
+	if err = cli.Exec(); err != nil {
+		panic(fmt.Sprintf("An error occurred when starting maintenance of the '%s': '%s'. ",
+			env.Vars.SystemName,
+			err))
+	}
+}