+++ src/internal/system/init_cli/config.go
@@ -1,0 +1,54 @@
+package init_cli
+
+import (
+	"sm-box/pkg/core/components/configurator"
+	"sm-box/pkg/core/components/tracer"
+)
+
+var confProfile = configurator.PrivateProfile{
+	Dir:      "/system/",
+	Filename: "init-cli.xml",
+}
+
+// Config - конфигурация CLI.
+type Config struct {
+}
+
+// FillEmptyFields - заполнение пустых полей конфигурации
+func (conf *Config) FillEmptyFields() *Config {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.FunctionCallFinished(conf) }()
+	}
+
+	return conf
+}
+
+// Default - запись стандартной конфигурации.
+func (conf *Config) Default() *Config {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.FunctionCallFinished(conf) }()
+	}
+
+	return conf
+}
+
+// Validate - валидация конфигурации.
+func (conf *Config) Validate() (err error) {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelConfig)
+
+		trc.FunctionCall()
+		defer func() { trc.Error(err).FunctionCallFinished() }()
+	}
+
+	return
+}