+++ src/scripts/build_win64.bash
@@ -1,5 +1,0 @@
-go env -w GOOS=windows
-
-go build -o F:/projects/SweetMint/sm-box/box/bin/box.exe ./cmd/app
-
-go build -o F:/projects/SweetMint/sm-box/box/sbin/init.exe ./cmd/system/init_script